requirements.txt
Flask
Flask-Session
google-auth
google-auth-oauthlib
google-auth-httplib2
google-auth-transport-requests
google-api-python-client
requests
pandas
gunicorn
